ai_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch
torchvision
Pillow
opencv-python
numpy
